search & chat
made it such that you search for a user's username / email and it shows all their posts.

made it such that when you search for a user and go to their profile, you can send them a message. (still working on the chat tab at the bottom for now. will get back to it soon)